+++ CHANGELOG.md
@@ -3,3 +3,10 @@ All notable changes to this project will be documented in this file.
 `SJSegmentedScrollView` adheres to [Semantic Versioning](http://semver.org/).
 
 #### 1.x Releases
+
+- `1.0.0` to `1.0.10`
+* Initial releases
+
+- `1.0.11`
+* Added objective c project support
+* Resolved header view height nil issue